requirements.txt
flask
requests
groq
python-dotenv
sqlite-utils
